redux file structure
