+++ client/src/pages/Tasks.tsx
@@ -3,8 +3,8 @@ import TaskPageHeader from "../components/Tasks/TaskPageHeader";
 import { useState, useEffect } from "react";
 import { Task } from "../interfaces/Task";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchAllTasks } from "../store/TaskDataSlice"; // Update the path
-import { RootState, AppDispatch } from "../store/store"; // Make sure to import your store types
+import { fetchAllTasks } from "../redux/TaskDataSlice"; // Update the path
+import { RootState, AppDispatch } from "../redux/store"; // Make sure to import your store types
 
 export default function Tasks() {
   const dispatch = useDispatch<AppDispatch>();
